JAMES-3621 Move JPA-SMTP mariaDB tests to main JPA-SMTP package
